refactor: Button component refactored
Refactored to be inline with the coding guidelines.
