+++ src/login/LoginOnboarding.tsx
@@ -3,7 +3,7 @@ import {StyleSheet} from '@atb/theme';
 import {LoginTexts, useTranslation} from '@atb/translations';
 import React from 'react';
 import {ScrollView, TouchableOpacity, View} from 'react-native';
-import Button from '@atb/components/button';
+import {Button} from '@atb/components/button';
 import {ThemeText} from '@atb/components/text';
 import {ArrowRight} from '@atb/assets/svg/mono-icons/navigation';
 import {LeftButtonProps, RightButtonProps} from '@atb/components/screen-header';
