+++ src/login/PhoneInput.tsx
@@ -9,7 +9,7 @@ import {
   View,
 } from 'react-native';
 import * as Sections from '@atb/components/sections';
-import Button from '@atb/components/button';
+import {Button} from '@atb/components/button';
 import {useAuthState} from '@atb/auth';
 import {ThemeText} from '@atb/components/text';
 import {PhoneSignInErrorCode} from '@atb/auth/AuthContext';
