+++ src/screens/Nearby/DepartureTimeSheet.tsx
@@ -1,6 +1,6 @@
 import {Confirm} from '@atb/assets/svg/mono-icons/actions';
 import {BottomSheetContainer} from '@atb/components/bottom-sheet';
-import Button from '@atb/components/button';
+import {Button} from '@atb/components/button';
 import FullScreenFooter from '@atb/components/screen-footer/full-footer';
 import {ScreenHeaderWithoutNavigation} from '@atb/components/screen-header';
 import {DateInputItem, Section, TimeInputItem} from '@atb/components/sections';
