+++ src/screens/Onboarding/WelcomeScreen.tsx
@@ -1,5 +1,5 @@
 import {Onboarding1} from '@atb/assets/svg/color/images/';
-import Button from '@atb/components/button';
+import {Button} from '@atb/components/button';
 import {ThemeText} from '@atb/components/text';
 import {StyleSheet} from '@atb/theme';
 import {StaticColorByType} from '@atb/theme/colors';
