+++ src/screens/Ticketing/FareContracts/PurchaseReservation.tsx
@@ -1,4 +1,4 @@
-import Button from '@atb/components/button';
+import {Button} from '@atb/components/button';
 import {ThemeText} from '@atb/components/text';
 import {StyleSheet, useTheme} from '@atb/theme';
 import {Reservation, PaymentType} from '@atb/ticketing';
