+++ src/screens/Ticketing/Purchase/Payment/Vipps/index.tsx
@@ -1,5 +1,5 @@
 import {ErrorType} from '@atb/api/utils';
-import Button from '@atb/components/button';
+import {Button} from '@atb/components/button';
 import {MessageBox} from '@atb/components/message-box';
 import FullScreenHeader from '@atb/components/screen-header/full-header';
 import {StyleSheet} from '@atb/theme';
