+++ src/screens/TripDetails/components/PaginatedDetailsHeader.tsx
@@ -4,7 +4,7 @@ import {PaginationTexts, useTranslation} from '@atb/translations';
 import {fullDateTime} from '@atb/utils/date';
 import React from 'react';
 import {View, ViewProps} from 'react-native';
-import Button from '@atb/components/button';
+import {Button} from '@atb/components/button';
 import {ThemeText} from '@atb/components/text';
 
 type PaginatedDetailsHeader = ViewProps & {
