+++ src/service-disruptions/ServiceDisruptionSheet.tsx
@@ -1,5 +1,5 @@
 import {BottomSheetContainer} from '@atb/components/bottom-sheet';
-import Button from '@atb/components/button';
+import {Button} from '@atb/components/button';
 import FullScreenFooter from '@atb/components/screen-footer/full-footer';
 import {ScreenHeaderWithoutNavigation} from '@atb/components/screen-header';
 import * as Sections from '@atb/components/sections';
